Create chapel_indexer.chpl
Certainly, we can create a basic file indexer in Chapel. The program will traverse the specified directory and print out the names of the files. If a file can't be opened, it will be noted as such.

Here's an example of how you might start creating such a program in Chapel:

In this example, the config const path allows you to set the directory you want to index from the command line with a --path argument, otherwise it defaults to the current directory ('.'). The indexFiles procedure opens each directory it encounters and recursively traverses any subdirectories.

Please make sure you have Chapel installed and correctly configured to compile and run the code. Chapel is a relatively new language and might not be present on some systems by default.