Create chapel_indexer_CLI.chpl
Directory indexer.

Sure, we can modify the program to prompt the user for the directory path and the output file name. We'll write the directory listing to this file.

Here's the revised Chapel code:

This version of the program will prompt the user to enter the path to the directory they want to index and the name of the output file. The indexFiles function now accepts a writer channel as a second argument, and writes the directory listing to this file. It also handles errors by writing an error message to the output file if it can't open a directory.