Create distributed_word_countv2.chpl
This program is a distributed version of the classic word count problem. The program will read a large text file, distributed across multiple nodes, and count the number of occurrences of each word in parallel. Be sure to include user prompts for ease of use in a command line environment.

This is a more complex problem because Chapel does not directly support distributed I/O. Instead, we need to read the file on one locale, then distribute the data to all locales. We'll also leverage Map for word counting. Here's a simple implementation:

To use this program, you would compile it with the Chapel compiler, then run the resulting binary, passing the path to the input file as a command-line argument:

chpl -o wordcount distributed_word_count.chpl
./wordcount --filePath=text.txt

This program reads the input file, breaks it into words, then distributes the words across all locales. It then uses a distributed map to count the occurrences of each word in parallel. Finally, it sorts the keys of the map (the words) and prints the counts for each word.

This is a basic implementation and doesn't handle all edge cases. In a production code, you would want to add error checking and handle a wider range of input data.