Create parallel_sorting.chpl
This program  performs sorting on a large set of data using a parallel sorting algorithm. The data is distributed across the available nodes and cores, and the sorting is performed in parallel. 

In Chapel, there are a variety of ways to implement parallel sorting. One common approach is to use the sort routine provided by the Sort module, which supports parallel sorting of distributed arrays. Below is an example of how this might be done:

This program reads data from a file into a distributed array, then sorts the array in parallel using Chapel's built-in sort function. After sorting, it prints out the sorted data. If the data contains non-numeric values (checked by isnan function), it returns an error message and stops execution.

You can compile and run this program with the Chapel compiler and pass in the number of elements and the file path as command-line arguments:

chpl -o parallel_sorting parallel_sorting.chpl
./parallel_sorting --numElements=1000000 --filePath=data.txt

Please note that this is a basic implementation and there's room for optimization. For example, you could use a parallel I/O approach to read the data file in parallel across multiple nodes.